requirements.txt
# Core ML/AI Libraries
torch>=2.1.0
transformers>=4.36.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.7.0
trl>=0.7.0
bitsandbytes>=0.41.0

# RAG and Vector Database
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
chromadb>=0.4.15
langchain>=0.0.350
langchain-community>=0.0.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
nltk>=3.8.0
spacy>=3.7.0

# API and Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0
httpx>=0.25.0

# Monitoring and Logging
wandb>=0.16.0
tensorboard>=2.15.0
loguru>=0.7.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.66.0
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Additional dependencies for document processing
PyPDF2>=3.0.1
python-docx>=0.8.11
python-multipart>=0.0.6

# Web scraping dependencies  
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium
webdriver-manager
scrapy-playwright

# OpenAI API integration
openai>=1.0.0

# Development Tools
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0
pre-commit>=3.5.0

